synthetic_proposals/requirements.txt
Faker>=19.0.0,<20.0.0
PyYAML>=6.0.0,<7.0.0
scikit-learn>=1.4.0,<1.6.0
torch>=2.0.0,<3.0.0
transformers>=4.30.0,<5.0.0
datasets>=2.14.0,<3.0.0
accelerate>=0.20.0,<1.0.0
